% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dori_b.R
\name{dori_b}
\alias{dori_b}
\title{Balanced DORI (vectorized over rows)}
\usage{
dori_b(groups, vars)
}
\arguments{
\item{groups}{matrix with rows = units, columns = categories}

\item{vars}{vector assigning each column to a variable (e.g., c(1,1,2,2))}
}
\value{
numeric vector of DORI_b values
}
\description{
Balanced DORI (vectorized over rows)
}
