% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dori_w.R
\name{dori_w}
\alias{dori_w}
\title{Weighted DORI (fully automatic, user-friendly)}
\usage{
dori_w(groups, groups_soc, vars)
}
\arguments{
\item{groups}{matrix of legislature proportions (one row per unit)}

\item{groups_soc}{matrix of population proportions (one row per unit)}

\item{vars}{variable assignment, e.g. c(1,1,2,2)}
}
\value{
numeric vector of DORI_w values
}
\description{
Weighted DORI (fully automatic, user-friendly)
}
